fix: Go files could not find the vuln lines in the files
Refs: K9VULN-7505

fix: lines not found again, but solved another way to make it more systematic

Refs: K9VULN-7505
