+++ assets/queries/cicd/github/run_block_injection/test/positive1.yaml
@@ -27,7 +27,7 @@ jobs:
       - name: Crawl pages and generate Markdown files
         uses: freeCodeCamp-China/article-webpage-to-markdown-action@v0.1.8
         with:
-          newsLink: '${{ github.event.issue.Body }}'
+          newsLink: '${{ github.event.issue.body }}'
           markDownFilePath: './chinese/articles/'
           githubToken: ${{ github.token }}
       - name: Git Auto Commit
