+++ assets/queries/cicd/github/unpinned_actions_full_length_commit_sha/query.rego
@@ -11,7 +11,7 @@ CxPolicy[result] {
 	
 	result := {
 		"documentId": input.document[i].id,
-		"searchKey": sprintf("uses=%s", [uses]),
+		"searchKey": sprintf("uses={{%s}}", [uses]),
 		"issueType": "IncorrectValue",
 		"keyExpectedValue": "Action pinned to a full length commit SHA.",
 		"keyActualValue": "Action is not pinned to a full length commit SHA.",
